+++ app/controllers/services_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class ServicesController < ApplicationController
+  EXTERNAL_SEARCH_ENABLED = Mp::Application.config.enable_external_search
+
   include Service::Searchable
   include Service::Categorable
   include Service::Autocomplete
@@ -14,8 +16,7 @@ class ServicesController < ApplicationController
   # rubocop:disable Metrics/AbcSize
   def index
     search_base_url = Mp::Application.config.search_service_base_url
-    enable_external_search = Mp::Application.config.enable_external_search
-    redirect_to search_base_url + "/search/service?q=*" if enable_external_search
+    redirect_to search_base_url + "/search/service?q=*" if external_search_enabled
 
     if params["object_id"].present?
       case params["type"]
@@ -134,5 +135,9 @@ class ServicesController < ApplicationController
     else
       []
     end
+  end
+
+  def external_search_enabled
+    EXTERNAL_SEARCH_ENABLED
   end
 end